Fix potential go routine leak
Feeds, FeedsByChat and Subs are functions that return a channel to read from.
This channel is filled with data from the database using an active connection.
If the code using this channel is not reading all of it, the go routine gets
stuck at the send and leaks, together with the channel and the db connection.
To fix this, the channel send is now in a select statement with the ctx.Done()
channel of the context passed into the function. Assuming the context is
cancelled soon after stopping the read from the channel, this is fine.
It is also good from a correctness perspective as the go routine is
maintaining the database connection and that should be closed when the given
context expires.